Create README.md
Climate Challenge is an initiative that presents a distinctive opportunity for tech-savvy individuals to collaborate with experts and create pioneering data-driven solutions for the urgent climate challenges faced by agriculture and fisheries sectors. This Climate challenge process is designed as a hackathon, to harness the power of data in developing impactful and sustainable solutions. Most importantly, applicants will have the opportunity to access investment funding to implement and scale up their innovative solutions.